requirements.txt
PyGithub
requests
pendulum
rich
openai
python-dotenv
BingImageCreator==0.4.4
todoist-api-python